fix: integrate error handling with extracted CLI program
Move error handling from cli.ts into a shared runProgram() function in
cli-program.ts so both the CLI entry point and integration test harness
use the same error handling logic.
